+++ test/CMakeLists.txt
@@ -23,8 +23,8 @@ find_package(Boost 1.67.0 REQUIRED COMPONENTS unit_test_framework)
 add_executable(ArborX_AccessTraits.exe tstAccessTraits.cpp)
 target_link_libraries(ArborX_AccessTraits.exe PRIVATE ArborX)
 
-add_executable(ArborX_DetailsCallbacks.exe tstDetailsCallbacks.cpp)
-target_link_libraries(ArborX_DetailsCallbacks.exe PRIVATE ArborX)
+add_executable(ArborX_Callbacks.exe tstCallbacks.cpp)
+target_link_libraries(ArborX_Callbacks.exe PRIVATE ArborX)
 
 add_executable(ArborX_DetailsConcepts.exe tstDetailsConcepts.cpp)
 target_link_libraries(ArborX_DetailsConcepts.exe PRIVATE ArborX)
